backend/requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.11
about-time==3.1.1
aiohttp==3.8.1
aiosignal==1.2.0
alive-progress==2.3.1
async-timeout==4.0.2
asynctest==0.13.0
attrs==21.4.0
Django==3.2
django-cors-headers==3.11.0
django-filter==21.1
djangorestframework==3.13.1
djangorestframework-camel-case==1.3.0
djangorestframework-simplejwt==5.1.0
frozenlist==1.3.0
grapheme==0.6.0
idna==3.3
importlib-metadata==4.11.2
Markdown==3.3.6
multidict==6.0.2
Pillow==9.0.0
progress==1.6
pydotplus==2.0.2
PyJWT==2.3.0
pyparsing==3.0.6
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
requests==2.27.1
wakeonlan==2.1.0
six==1.16.0
sqlparse==0.4.2
typing_extensions==4.0.1
urllib3==1.26.8
yarl==1.7.2
zipp==3.7.0
psycopg2-binary==2.9.3
gunicorn
